+++ core/beget.go
@@ -19,8 +19,9 @@ type Client struct {
 	Password string // Password for API access of Beget.com user
 }
 
-// PrepareRequest is a function to prepare request to Beget.API system.
-// As input parameters, it expects to receive information about the user and a built api method to call.
+// PrepareRequestWithClient is a function to prepare request to Beget.API system.
+// As input parameters, it expects to receive information about the user and a built api method to call and http.Client.
+// If http.client is nil, it will be created as an empty http.Client.
 // As a result, either a ready-to-execute request to the Beget.API is returned, or one of the errors:
 //   - Errors creating api method.
 //   - Errors resolving BEGET host
@@ -37,14 +38,19 @@ type Client struct {
 //		req, _ := PrepareRequest[result.GetData](client, dns.CallGetData("some.domain.com"))
 //
 // [GetData]: https://beget.com/ru/kb/api/funkczii-upravleniya-dns#getdata
-func PrepareRequest[Result any](c Client, method APIMethod[Result]) (*BegetRequest[Result], error) {
+func PrepareRequestWithClient[Result any](c Client, method APIMethod[Result], client *http.Client) (*BegetRequest[Result], error) {
 	if method.Error() != nil {
 		return nil, errors.Wrap(method.Error(), ErrFromAPIMethod.Error())
 	}
 
 	apiURL := method.GetURL()
 
-	hostURL, err := url.Parse(Host)
+	host := Host
+	if GetMode() == Test {
+		host = testHost
+	}
+
+	hostURL, err := url.Parse(host)
 	if err != nil {
 		return nil, errors.Wrap(err, ErrResolveBegetHost.Error())
 	}
@@ -66,5 +72,11 @@ func PrepareRequest[Result any](c Client, method APIMethod[Result]) (*BegetReque
 	// Add the user agent to the request.
 	httpRequest.Header.Add("User-Agent", UserAgent)
 
-	return NewBegetRequest[Result](httpRequest), nil
+	return NewBegetRequest[Result](httpRequest, client), nil
+}
+
+// PrepareRequest is a function to prepare request to Beget.API system.
+// The function is an alias in the PrepareRequestWithClient for working with the default http client.
+func PrepareRequest[Result any](c Client, method APIMethod[Result]) (*BegetRequest[Result], error) {
+	return PrepareRequestWithClient[Result](c, method, nil)
 }