+++ core/request.go
@@ -16,10 +16,13 @@ type BegetRequest[Result any] struct {
 }
 
 // NewBegetRequest creates a new [BegetRequest].
-func NewBegetRequest[Result any](req *http.Request) *BegetRequest[Result] {
+func NewBegetRequest[Result any](req *http.Request, client *http.Client) *BegetRequest[Result] {
+	if client == nil {
+		client = http.DefaultClient
+	}
 	return &BegetRequest[Result]{
 		req:    req,
-		client: http.DefaultClient,
+		client: client,
 	}
 }